requirements.txt
-i https://pypi.python.org/simple
asgiref==3.4.1; python_version >= '3.6'
backports.zoneinfo==0.2.1; python_version < '3.9'
boto3==1.17.49
botocore==1.20.112; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
celery ==5.0.2
certifi==2021.10.8
charset-normalizer==2.0.10; python_version >= '3'
decorator==4.4.2
django-cors-headers==3.7.0
django-crontab==0.7.1
django-cte==1.2.1
django-debug-toolbar-request-history==0.1.1
django-debug-toolbar==2.2
django-paypal==1.1.1
django-storages==1.11.1
django==3.2
djangorestframework==3.12.4
flake8==4.0.1
idna==3.3; python_version >= '3'
imageio-ffmpeg==0.4.5; python_version >= '3.4'
imageio==2.14.0; python_version >= '3.5'
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
mccabe==0.6.1
moviepy==1.0.3
numpy==1.22.1; python_version >= '3.8'
packaging==21.3
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
pillow==8.3.2
proglog==0.1.9
psycopg2==2.9.1
pycodestyle==2.8.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pyfcm==1.5.3
pyflakes==2.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.3
requests==2.27.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
s3transfer==0.3.7
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
tqdm==4.62.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
#pycurl==7.43.0.5
django-environ==0.9.0
